Log in From && Student Controller Redirect

+++ Ex3_ASPNETWebAppMVCStudentApp/Controllers/StudentsController.cs
@@ -13,40 +13,53 @@ namespace Ex3_ASPNETWebAppMVCStudentApp.Controllers
     {
         private Lab6_Ex3Entities db = new Lab6_Ex3Entities();
 
+        private bool IsLoggedIn()
+        {
+            return Session["Username"] != null;
+        }
+
         // GET: Students
         public ActionResult Index()
         {
+            if (!IsLoggedIn())
+                return RedirectToAction("Login", "Account");
+
             return View(db.Students.ToList());
         }
 
         // GET: Students/Details/5
         public ActionResult Details(int? id)
         {
+            if (!IsLoggedIn())
+                return RedirectToAction("Login", "Account");
+
             if (id == null)
-            {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
+
             Student student = db.Students.Find(id);
             if (student == null)
-            {
                 return HttpNotFound();
-            }
+
             return View(student);
         }
 
         // GET: Students/Create
         public ActionResult Create()
         {
+            if (!IsLoggedIn())
+                return RedirectToAction("Login", "Account");
+
             return View();
         }
 
         // POST: Students/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "StudentID,FirstName,LastName,EnrollmentDate")] Student student)
         {
+            if (!IsLoggedIn())
+                return RedirectToAction("Login", "Account");
+
             if (ModelState.IsValid)
             {
                 db.Students.Add(student);
@@ -60,25 +73,27 @@ namespace Ex3_ASPNETWebAppMVCStudentApp.Controllers
         // GET: Students/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (!IsLoggedIn())
+                return RedirectToAction("Login", "Account");
+
             if (id == null)
-            {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
+
             Student student = db.Students.Find(id);
             if (student == null)
-            {
                 return HttpNotFound();
-            }
+
             return View(student);
         }
 
         // POST: Students/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for 
-        // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "StudentID,FirstName,LastName,EnrollmentDate")] Student student)
         {
+            if (!IsLoggedIn())
+                return RedirectToAction("Login", "Account");
+
             if (ModelState.IsValid)
             {
                 db.Entry(student).State = EntityState.Modified;
@@ -91,15 +106,16 @@ namespace Ex3_ASPNETWebAppMVCStudentApp.Controllers
         // GET: Students/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (!IsLoggedIn())
+                return RedirectToAction("Login", "Account");
+
             if (id == null)
-            {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
+
             Student student = db.Students.Find(id);
             if (student == null)
-            {
                 return HttpNotFound();
-            }
+
             return View(student);
         }
 
@@ -108,6 +124,9 @@ namespace Ex3_ASPNETWebAppMVCStudentApp.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (!IsLoggedIn())
+                return RedirectToAction("Login", "Account");
+
             Student student = db.Students.Find(id);
             db.Students.Remove(student);
             db.SaveChanges();
